+++ SW4STM32/BellaLui/Application/HostBoard/Src/Sensors/FastI2CDriver.cpp
@@ -12,7 +12,7 @@
 
 FastI2CDriver::FastI2CDriver() : I2CDriver(&FastI2CDriver::read, &FastI2CDriver::write) {}
 
-int8_t FastI2CDriver::read(uint8_t dev_id, uint8_t reg_addr, uint8_t *data, uint16_t len) {
+int8_t FastI2CDriver::read(uint8_t dev_id, uint8_t reg_addr, uint8_t *data, uint8_t len) {
 	int8_t rslt = 0;
 
 	vTaskSuspendAll();
@@ -24,7 +24,7 @@ int8_t FastI2CDriver::read(uint8_t dev_id, uint8_t reg_addr, uint8_t *data, uint
 	return rslt;
 }
 
-int8_t FastI2CDriver::write(uint8_t dev_id, uint8_t reg_addr, uint8_t *data, uint16_t len) {
+int8_t FastI2CDriver::write(uint8_t dev_id, uint8_t reg_addr, uint8_t *data, uint8_t len) {
 	int8_t rslt = 0;
 
 	vTaskSuspendAll();
